services/ai-engine-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
python-dotenv==1.0.0
